+++ pkg/apis/kubermatic/v1/allowed_registry.go
