+++ pkg/apis/kubermatic/v1/cluster_template_instance.go
