+++ pkg/apis/kubermatic/v1/cluster_templates.go
