+++ pkg/apis/kubermatic/v1/constraint_template.go
@@ -17,7 +17,7 @@ limitations under the License.
 package v1
 
 import (
-	openpolicyagent "k8c.io/api/v2/pkg/open-policy-agent"
+	openpolicyagent "k8c.io/api/v2/pkg/apis/open-policy-agent"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
