+++ pkg/apis/kubermatic/v1/external_cluster.go
@@ -17,9 +17,8 @@ limitations under the License.
 package v1
 
 import (
-	machinecontroller "k8c.io/api/v2/pkg/machine-controller"
+	machinecontroller "k8c.io/api/v2/pkg/apis/machine-controller"
 	"k8c.io/api/v2/pkg/semver"
-	"k8c.io/api/v2/pkg/types"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -132,7 +131,7 @@ type ExternalClusterNetworkRanges struct {
 
 // ExternalClusterCloudSpec mutually stores access data to a cloud provider.
 type ExternalClusterCloudSpec struct {
-	ProviderName types.ExternalClusterProvider         `json:"providerName"`
+	ProviderName ExternalClusterProvider               `json:"providerName"`
 	GKE          *ExternalClusterGKECloudSpec          `json:"gke,omitempty"`
 	EKS          *ExternalClusterEKSCloudSpec          `json:"eks,omitempty"`
 	AKS          *ExternalClusterAKSCloudSpec          `json:"aks,omitempty"`
