+++ pkg/apis/kubermatic/v1/helper/datacenter_test.go
@@ -19,7 +19,7 @@ package helper
 import (
 	"testing"
 
-	kubermaticv1 "k8c.io/api/v2/pkg/kubermatic/v1"
+	kubermaticv1 "k8c.io/api/v2/pkg/apis/kubermatic/v1"
 	"k8c.io/api/v2/pkg/test/diff"
 
 	"k8s.io/utils/pointer"
