+++ pkg/apis/kubermatic/v1/helper/helper_test.go
@@ -19,7 +19,7 @@ package helper
 import (
 	"testing"
 
-	kubermaticv1 "k8c.io/api/v2/pkg/kubermatic/v1"
+	kubermaticv1 "k8c.io/api/v2/pkg/apis/kubermatic/v1"
 
 	corev1 "k8s.io/api/core/v1"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
