+++ pkg/apis/kubermatic/v1/zz_generated.deepcopy.go
@@ -23,10 +23,9 @@ package v1
 
 import (
 	"encoding/json"
-	machine_controller "k8c.io/api/v2/pkg/machine-controller"
-	open_policy_agent "k8c.io/api/v2/pkg/open-policy-agent"
+	machine_controller "k8c.io/api/v2/pkg/apis/machine-controller"
+	open_policy_agent "k8c.io/api/v2/pkg/apis/open-policy-agent"
 	"k8c.io/api/v2/pkg/semver"
-	"k8c.io/api/v2/pkg/types"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -4287,7 +4286,7 @@ func (in *KubermaticVersioningConfiguration) DeepCopyInto(out *KubermaticVersion
 	}
 	if in.ExternalClusters != nil {
 		in, out := &in.ExternalClusters, &out.ExternalClusters
-		*out = make(map[types.ExternalClusterProvider]ExternalClusterProviderVersioningConfiguration, len(*in))
+		*out = make(map[ExternalClusterProvider]ExternalClusterProviderVersioningConfiguration, len(*in))
 		for key, val := range *in {
 			(*out)[key] = *val.DeepCopy()
 		}
